+++ src/components/message/message-test.stories.tsx
@@ -1,10 +1,11 @@
-import React, { useState } from "react";
+import React, { useState, useRef, useEffect } from "react";
 import { action } from "@storybook/addon-actions";
 import Button from "../button";
 import Message, { MessageProps } from "./message.component";
 
 export default {
   title: "Message/Test",
+  includeStories: ["Default"],
   parameters: {
     info: { disable: true },
     chromatic: {
@@ -61,6 +62,29 @@ export const MessageComponent = (props: MessageProps) => {
     <div>
       {!isOpen && <Button onClick={() => setIsOpen(true)}>Open Message</Button>}
       <Message open={isOpen} onDismiss={() => setIsOpen(false)} {...props}>
+        Some custom message
+      </Message>
+    </div>
+  );
+};
+
+export const MessageComponentWithRef = (props: MessageProps) => {
+  const [isOpen, setIsOpen] = useState(false);
+  const messageRef: React.Ref<HTMLDivElement> = useRef(null);
+
+  useEffect(() => {
+    if (isOpen) messageRef.current?.focus();
+  });
+
+  return (
+    <div>
+      {!isOpen && <Button onClick={() => setIsOpen(true)}>Open Message</Button>}
+      <Message
+        open={isOpen}
+        onDismiss={() => setIsOpen(false)}
+        ref={messageRef}
+        {...props}
+      >
         Some custom message
       </Message>
     </div>